qwerty

+++ sketch.js
@@ -2,7 +2,6 @@
 // https://stackoverflow.com/questions/61747298/make-a-calculator-in-p5
 // https://stackoverflow.com/questions/61757326/making-a-calculator-function-properly-in-p5-js
 var mouseClick = false;
-//testtettetsetetswfajdiaeofjeaijfeaiojdiaojdioajd
 
 var displayNumArray = [];  //gloab array er nødvendig for at få tal fra flere objecter
 var operation;
@@ -72,7 +71,7 @@ function draw()
   {
     displayNum += displayNumArray[i]; //indsæt ud fra rækkefølge
   }
-  wholeNum = parseInt(displayNum); //vi har en seperat variabel som vi kan bruge til udregning (display num er kun til at vise tallet)
+  wholeNum = parseFloat(displayNum); //vi har en seperat variabel som vi kan bruge til udregning (display num er kun til at vise tallet)
   
   if(!isNaN(wholeNum))
   {
